Skip muted list update if no change
`selectMutedChannels` run and return a new reference each time the app gains focus, causing components like `ClaimPreview` and several others to render unnecessarily.

`createSelector` will not help in this case, because we are indeed invalidating the data in the store.

Not sure if it's a valid pattern to return existing state, but it helps in not invalidating cache.