Fix sidebar appearing on livestreams
## Ticket
75 Stream video is overtaking chat due to sidebar

## Issue

- https://odysee.com/@TurdFlingingMonkey#1/TFMShowLIVEOdyseeStream:7
- https://odysee.com/@TurdFlingingMonkey:1/TFMShowLIVEOdyseeStream:7

When the # format is used for the channel name, `parseURI` thinks it's a channel page.

## Fix
Since we have a `livestream` flag, use that to correct the logic. This is the more localized fix that requires minimal testing.

The actual fix would be at the `parseURI` side -- submitting a separate issue for that.