Handle quick-actions incorrectly reverting previous change
## Repro
1. Follow a channel.
2. When `preference_set` is sent, unfollow the channel.
3. A few seconds later, the final setting reflects (1) instead of (2).

The current sync loop involves doing a final `sync/get` at the end. While not necessary for the scenario above, the code flow covers various cases, so it's still needed for now.

## Approach
When syncing from the `buildSharedStateMiddleware` loop, generate an ID for each session, and store the latest one. Pass the ID all the way to the completion-callback, so we can check if our session is still the latest one. If not, there is no need to do the final `sync/get`.

The check for invalidation can also be placed in more places to cut off the process earlier, but it's only done for 2 critical places for now.