Handle syncing while the token is being refreshed
Ignoring the error for now, i.e. not mark it as 'fatal error' which will invoke the Nag. The next sync cycle will fix things.

If this is not good enough, alternatives:

(1) use `keycloak.isTokenExpired(N)` to detect if it will be refreshing in N second and figure out how to delay `doGetSync`. This requires moving things around, which I prefer not to do since this solution isn't perfect either (i.e. requires arbitrary delay).

(2) We can call sync on every `ACTIONS.OAUTH_REFRESH_SUCCESS` to ensure nothing is missed (although I still think it's not necessary).