notify: remove forced timeout, just beautify message
## Issue
The forced timeout seems to be causing more errors than usual on large files, which naturally take longer on the SDK side.

## Change
Just beautify the message instead of forcing a timeout. Should have done this in the first place.

Note that I wasn't able to fully test this given it's hard to replicate, but it shouldn't cause any problems even if the code fails to capture the scenario -- it will just show the original error message, and the "remove" modal will still explain things.