Don't show zero stats while fetching
## Issue
Sub count, view count and reactions are incorrectly showing "0" briefly, or stuck there if the fetch fails.

## Reason
Selectors defaulting to 0.

## Change
(1) Apply placeholders or simply not show it (depending on which looks best for each scenario).

(2) To reduce layout shifts, swapped the location of "#uploads" vs "# followers" since the former will always be populated.