commentCreate: re-organize to make memoization easier
Putting an early jsx-return in the middle makes it harder to use `useMemo` and `useCallback` that requires a variable after the return call.

Try to follow this structure:
- Props destructured
- States and refs
- const calculations
- Functions
- Effects
- Return

handleStickerComment and handleStickerComment was moved upwards to make the next commit's diff cleaner. No code changes here other than changing it from a function to const