commentCreate: prevent CommentSelectors from re-rendering on input change (#1964)
* commentCreate: re-organize to make memoization easier

Putting an early jsx-return in the middle makes it harder to use `useMemo` and `useCallback` that requires a variable after the return call.

Try to follow this structure:
- Props destructured
- States and refs
- const calculations
- Functions
- Effects
- Return

handleStickerComment and handleStickerComment was moved upwards to make the next commit's diff cleaner. No code changes here other than changing it from a function to const

* commentCreate: update state based on actual prev rather than closure'd value

This is the recommended way to update a value anyways, but it's mainly to make memoization easier because we then don't need to put the variable in the dependency list.

* commentCreate: prevent CommentSelectors from re-rendering on input change

## Issue
Typing on the input field is slow (although ok in development build) because everything is re-rendered when `commentValue` changes. `EmoteCategory` is the slowest based on the profiler.

## Change
Memoize `<CommentSelectors>` so that `EmoteCategory` doesn't re-render.